--- frontend/guest-app/src/components/Question/QuestionsRepliesContext.js
@@ -1,4 +1,0 @@
-import {createContext} from "react";
-
-export const QuestionsRepliesContext = createContext([]);
-export const QuestionsRepliesProvider = QuestionsRepliesContext.Provider;
